Add a rough hack: support "socket_file:" section
The problem with config loading is that it's terribly complex and
slow. For loading socket configs we don't need all the complexity,
just want to load the basic flat file.

To enable that we introduce a proper hack: socket_file section,
which takes a single parameter: filename. Circus will load this
file and all socket: sections from it. No validation is done. This
is intended to be fast.